viz_project/requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
dnspython==2.4.2
mysql-connector-python==8.2.0
protobuf==4.21.12
pymongo==4.6.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
